Fix a condition where the algorithm would give the sunrise and sunset of the previous day in winter.
For some reason, tsouth would be computed with a negative value,
so the sunset and sunrise would occur during the eve and not the day.
Try with a date after 22 September or before 20 March.
And on the threshold date between the two periods, there would be a
flip-flop leading to a "not converging" condition.